Renamed generic unlocks namespace to singular form for consistency. Small style cleanup.
