+++ Source/NexusForever.Network.World/Message/Model/GenericUnlock/ClientItemGenericUnlock.cs
@@ -1,7 +1,7 @@
 using NexusForever.Network.Message;
 using NexusForever.Network.World.Message.Model.Shared;
 
-namespace NexusForever.Network.World.Message.Model.GenericUnlocks
+namespace NexusForever.Network.World.Message.Model.GenericUnlock
 {
     [Message(GameMessageOpcode.ClientItemGenericUnlock)]
     public class ClientItemGenericUnlock : IReadable
