+++ Source/NexusForever.Network.World/Message/Model/GenericUnlock/ServerGenericUnlockAccountList.cs
@@ -1,11 +1,11 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.GenericUnlocks
+namespace NexusForever.Network.World.Message.Model.GenericUnlock
 {
     [Message(GameMessageOpcode.ServerGenericUnlockAccountList)]
     public class ServerGenericUnlockAccountList : IWritable
     {
-        public List<uint> GenericUnlockEntryIds { get; set; } = new();
+        public List<uint> GenericUnlockEntryIds { get; set; } = [];
 
         public void Write(GamePacketWriter writer)
         {
