+++ Source/NexusForever.Network.World/Message/Model/GenericUnlock/ServerGenericUnlockCharacterList.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.GenericUnlocks
+namespace NexusForever.Network.World.Message.Model.GenericUnlock
 {
     // Sets bCharacterUnlocked state on all GenericUnlocks on the client.
     // Seems that ServerGenericUnlocksCharacterRefreshed was the preferred method
@@ -8,7 +8,7 @@ namespace NexusForever.Network.World.Message.Model.GenericUnlocks
     [Message(GameMessageOpcode.ServerGenericUnlockCharacterList)]
     public class ServerGenericUnlockCharacterList : IWritable
     {
-        public List<uint> GenericUnlockEntryIds { get; set; } = new();
+        public List<uint> GenericUnlockEntryIds { get; set; } = [];
 
         public void Write(GamePacketWriter writer)
         {
