+++ Source/NexusForever.Network.World/Message/Model/GenericUnlock/ServerGenericUnlockCharacterRefresh.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.GenericUnlocks
+namespace NexusForever.Network.World.Message.Model.GenericUnlock
 {
     // Clears the bCharacterUnlocked state on all GenericUnlocks on the client then
     // uses the provided array to set the state again. Seems to be used preferentially
@@ -8,7 +8,7 @@ namespace NexusForever.Network.World.Message.Model.GenericUnlocks
     [Message(GameMessageOpcode.ServerGenericUnlockCharacterRefresh)]
     public class ServerGenericUnlockCharacterRefresh : IWritable
     {
-        public List<uint> GenericUnlockEntryIds { get; set; } = new();
+        public List<uint> GenericUnlockEntryIds { get; set; } = [];
 
         public void Write(GamePacketWriter writer)
         {
