+++ Source/NexusForever.WorldServer/Network/Message/Handler/Item/ClientItemGenericUnlockHandler.cs
@@ -3,7 +3,7 @@ using NexusForever.GameTable;
 using NexusForever.GameTable.Model;
 using NexusForever.Network;
 using NexusForever.Network.Message;
-using NexusForever.Network.World.Message.Model.GenericUnlocks;
+using NexusForever.Network.World.Message.Model.GenericUnlock;
 
 namespace NexusForever.WorldServer.Network.Message.Handler.Item
 {
